feat: implement filter logic, refactor vacancies page, refactor code
