--- job-search/src/pages/VacanciesPage/VacanciesForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from './VacanciesForm'
-export * from './hooks'
