+++ job-search/src/pages/VacanciesPage/VacanciesPage.module.scss
@@ -8,6 +8,89 @@
   width: 100%;
 }
 
+.filterBlock {
+  background-color: var(--white-color);
+  border: 1px solid var(--grey-color-200);
+  border-radius: 12px;
+  display: flex;
+  flex-direction: column;
+  height: 360px;
+  padding: 19px;
+  width: 315px;
+}
+
+.titleBlock {
+  display: flex;
+  flex-direction: row;
+  gap: 67px;
+  width: 100%;
+}
+
+.title {
+  color: var(--black-color);
+  font: var(--font-h2);
+}
+
+.resetBtn {
+  align-items: center;
+  background: none;
+  border: none;
+  cursor: pointer;
+  display: flex;
+  flex-direction: row;
+  gap: 4px;
+  margin-bottom: 32px;
+}
+
+.resetText {
+  color: var(--grey-color-500);
+  font: var(--font-text-3);
+}
+
+.closeImg {
+  height: 16px;
+  width: 16px;
+}
+
+.industryBlock {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  margin-bottom: 20px;
+}
+
+.subtitle {
+  color: var(--black-color);
+  font: var(--font-h3);
+}
+
+.salaryBlock {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  margin-bottom: 20px;
+}
+
+.salaryWrapper {
+  position: relative;
+}
+
+.filterButton {
+  background-color: var(--light-blue-color);
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  color: var(--white-color);
+  font: var(--font-text-3);
+  height: 40px;
+  transition: all 0.5s ease-in-out;
+  width: 275px;
+
+  &:hover {
+    transform: scale(0.95);
+  }
+}
+
 .searchBlock {
   align-items: center;
   display: flex;
